refactor: rename cli sbom command to ingest; create modules per command and handler.
