+++ cli/src/bin/main.rs
@@ -6,7 +6,7 @@ async fn main() -> Result<(), Error> {
     let cli = Cli::parse();
     match &cli.command {
         Some(Commands::Enrich(enrich)) => commands::enrich::execute(enrich).await,
-        Some(Commands::Sbom(sbom)) => commands::sbom::execute(sbom).await,
+        Some(Commands::Ingest(ingest)) => commands::ingest::execute(ingest).await,
         Some(Commands::Analyze(analyze)) => commands::analyze::execute(analyze).await,
         _ => {
             println!("command not found");
