+++ cli/src/commands/mod.rs
@@ -1,6 +1,8 @@
-/// Contains the types and functions to support the Enrich Command.
+/// Contains the types and functions to support the `enrich` Command.
 pub mod enrich;
-/// Contains the types and functions to support the SBOM Command.
-pub mod sbom;
-/// Contains the type and functions to support SBOM analytics.
+
+/// Contains the types and functions to support the `ingest` Command.
+pub mod ingest;
+
+/// Contains the type and functions to support `analyze` Command.
 pub mod analyze;
